+++ src/Updaters.tsx
@@ -1,0 +1,22 @@
+import React from 'react'
+import { MulticallUpdater } from 'lib/state/multicall'
+import ApplicationUpdater from './state/application/updater'
+import ListsUpdater from './state/lists/updater'
+import LogsUpdater from './state/logs/updater'
+import TransactionUpdater from './state/transactions/updater'
+import UserUpdater from './state/user/updater'
+
+function Updaters() {
+  return (
+    <>
+      <ListsUpdater />
+      <UserUpdater />
+      <ApplicationUpdater />
+      <TransactionUpdater />
+      <MulticallUpdater />
+      <LogsUpdater />
+    </>
+  )
+}
+
+export default Updaters
